Move mcpwebhookvalidation to the right place
